requirements.txt
fyers-apiv3
pandas
numpy
requests
python-dotenv
gunicorn
flask
